requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
httpx==0.24.1
python-dotenv==0.19.0
jinja2==3.0.1
pydantic==1.8.2